refactor: Change data type for cost_per_night to SMALLINT for consistancy

+++ migrations/01_schema.sql
@@ -18,7 +18,7 @@ CREATE TABLE properties(
   description TEXT,
   thumbnail_photo_url VARCHAR(255) NOT NULL,
   cover_photo_url VARCHAR(255) NOT NULL,
-  cost_per_night INTEGER NOT NULL DEFAULT 0,
+  cost_per_night SMALLINT NOT NULL DEFAULT 0,
   parking_spaces SMALLINT NOT NULL DEFAULT 0,
   number_of_bathrooms SMALLINT NOT NULL DEFAULT 0,
   number_of_bedrooms SMALLINT NOT NULL DEFAULT 0,
